+++ src/main/java/net/ttddyy/dsproxy/listener/logging/CommonsQueryLoggingListener.java
@@ -14,6 +14,33 @@ public class CommonsQueryLoggingListener extends AbstractQueryLoggingListener {
     protected Log log = LogFactory.getLog(CommonsQueryLoggingListener.class);
     protected CommonsLogLevel logLevel = CommonsLogLevel.DEBUG; // default DEBUG
 
+
+    public CommonsQueryLoggingListener() {
+
+        // initialize logging condition that checks current log level
+        this.loggingCondition = new LoggingCondition() {
+            @Override
+            public boolean getAsBoolean() {
+                switch (logLevel) {
+                    case DEBUG:
+                        return log.isDebugEnabled();
+                    case ERROR:
+                        return log.isErrorEnabled();
+                    case FATAL:
+                        return log.isFatalEnabled();
+                    case INFO:
+                        return log.isInfoEnabled();
+                    case TRACE:
+                        return log.isTraceEnabled();
+                    case WARN:
+                        return log.isWarnEnabled();
+                }
+                return false;
+            }
+        };
+
+    }
+
     @Override
     protected void writeLog(String message) {
         CommonsLogUtils.writeLog(log, this.logLevel, message);