+++ src/main/java/net/ttddyy/dsproxy/listener/logging/SLF4JQueryLoggingListener.java
@@ -14,6 +14,27 @@ public class SLF4JQueryLoggingListener extends AbstractQueryLoggingListener {
     protected Logger logger = LoggerFactory.getLogger(SLF4JQueryLoggingListener.class);
     protected SLF4JLogLevel logLevel = SLF4JLogLevel.DEBUG; // default DEBUG
 
+    public SLF4JQueryLoggingListener() {
+        this.loggingCondition = new LoggingCondition() {
+            @Override
+            public boolean getAsBoolean() {
+                switch (logLevel) {
+                    case TRACE:
+                        return logger.isTraceEnabled();
+                    case DEBUG:
+                        return logger.isDebugEnabled();
+                    case INFO:
+                        return logger.isInfoEnabled();
+                    case WARN:
+                        return logger.isWarnEnabled();
+                    case ERROR:
+                        return logger.isErrorEnabled();
+                }
+                return false;
+            }
+        };
+    }
+
     @Override
     protected void writeLog(String message) {
         SLF4JLogUtils.writeLog(logger, this.logLevel, message);