+++ src/main/java/net/ttddyy/dsproxy/listener/logging/SystemOutQueryLoggingListener.java
@@ -7,6 +7,16 @@ package net.ttddyy.dsproxy.listener.logging;
  * @since 1.3
  */
 public class SystemOutQueryLoggingListener extends AbstractQueryLoggingListener {
+
+    public SystemOutQueryLoggingListener() {
+        this.loggingCondition = new LoggingCondition() {
+            @Override
+            public boolean getAsBoolean() {
+                return true;  // always perform logging
+            }
+        };
+    }
+
     @Override
     protected void writeLog(String message) {
         System.out.println(message);