+++ src/test/java/net/ttddyy/dsproxy/listener/logging/CommonsQueryLoggingListenerTest.java
@@ -1,9 +1,16 @@
 package net.ttddyy.dsproxy.listener.logging;
 
+import net.ttddyy.dsproxy.ExecutionInfo;
+import net.ttddyy.dsproxy.ExecutionInfoBuilder;
+import net.ttddyy.dsproxy.QueryInfo;
+import net.ttddyy.dsproxy.QueryInfoBuilder;
 import org.apache.commons.logging.impl.NoOpLog;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -46,4 +53,57 @@ public class CommonsQueryLoggingListenerTest {
         assertThat(name).as("Updated logger name").isEqualTo("my.logger");
     }
 
+
+    @Test
+    public void loggingCondition() {
+        CommonsQueryLoggingListener listener = new CommonsQueryLoggingListener();
+        InMemoryCommonsLog log = new InMemoryCommonsLog();
+        listener.setLog(log);
+
+        ExecutionInfo execInfo = ExecutionInfoBuilder.create().build();
+        List<QueryInfo> queryInfoList = new ArrayList<QueryInfo>();
+        queryInfoList.add(QueryInfoBuilder.create().query("select * ").build());
+
+        // listener writes to more serious level
+        listener.setLogLevel(CommonsLogLevel.DEBUG);
+        log.setEnabledLogLevel(CommonsLogLevel.TRACE);
+
+        listener.afterQuery(execInfo, queryInfoList);
+
+        assertThat(log.getTraceMessages()).isEmpty();
+        assertThat(log.getDebugMessages()).hasSize(1);
+        assertThat(log.getInfoMessages()).isEmpty();
+        assertThat(log.getWarnMessages()).isEmpty();
+        assertThat(log.getErrorMessages()).isEmpty();
+        assertThat(log.getFatalMessages()).isEmpty();
+
+        // listener writes to less serious level
+        log.reset();
+        listener.setLogLevel(CommonsLogLevel.TRACE);
+        log.setEnabledLogLevel(CommonsLogLevel.DEBUG);
+
+        listener.afterQuery(execInfo, queryInfoList);
+
+        assertThat(log.getTraceMessages()).isEmpty();
+        assertThat(log.getDebugMessages()).isEmpty();
+        assertThat(log.getInfoMessages()).isEmpty();
+        assertThat(log.getWarnMessages()).isEmpty();
+        assertThat(log.getErrorMessages()).isEmpty();
+        assertThat(log.getFatalMessages()).isEmpty();
+
+        // listener writes to same level
+        log.reset();
+        listener.setLogLevel(CommonsLogLevel.DEBUG);
+        log.setEnabledLogLevel(CommonsLogLevel.DEBUG);
+
+        listener.afterQuery(execInfo, queryInfoList);
+
+        assertThat(log.getTraceMessages()).isEmpty();
+        assertThat(log.getDebugMessages()).hasSize(1);
+        assertThat(log.getInfoMessages()).isEmpty();
+        assertThat(log.getWarnMessages()).isEmpty();
+        assertThat(log.getErrorMessages()).isEmpty();
+        assertThat(log.getFatalMessages()).isEmpty();
+
+    }
 }