+++ src/test/java/net/ttddyy/dsproxy/listener/logging/JULQueryLoggingListenerTest.java
@@ -1,6 +1,14 @@
 package net.ttddyy.dsproxy.listener.logging;
 
+import net.ttddyy.dsproxy.ExecutionInfo;
+import net.ttddyy.dsproxy.ExecutionInfoBuilder;
+import net.ttddyy.dsproxy.QueryInfo;
+import net.ttddyy.dsproxy.QueryInfoBuilder;
 import org.junit.Test;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.logging.Level;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -22,6 +30,62 @@ public class JULQueryLoggingListenerTest {
         JULQueryLoggingListener listener = new JULQueryLoggingListener();
         listener.setLogger("my.logger");
         assertThat(listener.getLogger().getName()).as("Updated logger name").isEqualTo("my.logger");
+    }
+
+    @Test
+    public void loggingCondition() {
+        JULQueryLoggingListener listener = new JULQueryLoggingListener();
+        InMemoryJULLogger logger = new InMemoryJULLogger();
+        listener.setLogger(logger);
+
+        ExecutionInfo execInfo = ExecutionInfoBuilder.create().build();
+        List<QueryInfo> queryInfoList = new ArrayList<QueryInfo>();
+        queryInfoList.add(QueryInfoBuilder.create().query("select * ").build());
+
+        // listener writes to more serious level
+        listener.setLogLevel(Level.WARNING);
+        logger.setLoggerLevel(Level.INFO);
+
+        listener.afterQuery(execInfo, queryInfoList);
+
+        assertThat(logger.getSevereMessages()).isEmpty();
+        assertThat(logger.getWarningMessages()).hasSize(1);
+        assertThat(logger.getInfoMessages()).isEmpty();
+        assertThat(logger.getConfigMessages()).isEmpty();
+        assertThat(logger.getFineMessages()).isEmpty();
+        assertThat(logger.getFinerMessages()).isEmpty();
+        assertThat(logger.getFinestMessages()).isEmpty();
+
+        // listener writes to less serious level
+        logger.reset();
+        listener.setLogLevel(Level.FINE);
+        logger.setLoggerLevel(Level.WARNING);
+
+        listener.afterQuery(execInfo, queryInfoList);
+
+        assertThat(logger.getSevereMessages()).isEmpty();
+        assertThat(logger.getWarningMessages()).isEmpty();
+        assertThat(logger.getInfoMessages()).isEmpty();
+        assertThat(logger.getConfigMessages()).isEmpty();
+        assertThat(logger.getFineMessages()).isEmpty();
+        assertThat(logger.getFinerMessages()).isEmpty();
+        assertThat(logger.getFinestMessages()).isEmpty();
+
+        // listener writes to same level
+        logger.reset();
+        listener.setLogLevel(Level.WARNING);
+        logger.setLoggerLevel(Level.WARNING);
+
+        listener.afterQuery(execInfo, queryInfoList);
+
+        assertThat(logger.getSevereMessages()).isEmpty();
+        assertThat(logger.getWarningMessages()).hasSize(1);
+        assertThat(logger.getInfoMessages()).isEmpty();
+        assertThat(logger.getConfigMessages()).isEmpty();
+        assertThat(logger.getFineMessages()).isEmpty();
+        assertThat(logger.getFinerMessages()).isEmpty();
+        assertThat(logger.getFinestMessages()).isEmpty();
+
     }
 
 }